Javadocs and bug fixes

+++ src/main/java/by/bsuir/hockeyshop/command/impl/DeleteOrderCommand.java
@@ -2,6 +2,8 @@ package by.bsuir.hockeyshop.command.impl;
 
 import by.bsuir.hockeyshop.command.ActionCommand;
 import by.bsuir.hockeyshop.command.CommandException;
+import by.bsuir.hockeyshop.entity.User;
+import by.bsuir.hockeyshop.entity.UserRole;
 import by.bsuir.hockeyshop.managers.ConfigurationManager;
 import by.bsuir.hockeyshop.managers.MessageManager;
 import by.bsuir.hockeyshop.service.OrderService;
@@ -20,6 +22,7 @@ public class DeleteOrderCommand implements ActionCommand {
     static final String ATTR_SUCCESS = "successMessage";
     static final String ATTR_ERROR = "errorMessage";
     static final String COMMAND_VIEW_ORDERS = "/controller?command=view_user_orders";
+    static final String COMMAND_VIEW_UNPAID_ORDERS = "/controller?command=view_all_orders&type=unpaid";
     static final String ATTR_MESSAGE_MANAGER = "messageManager";
     private static final OrderService ORDER_SERVICE = OrderServiceImpl.getInstance();
 
@@ -33,13 +36,23 @@ public class DeleteOrderCommand implements ActionCommand {
     public String execute(HttpServletRequest request) throws CommandException {
         String page;
         long orderId = Long.parseLong(request.getParameter(PARAM_ORDER_ID));
+        User user = (User)(request.getSession().getAttribute(ATTR_USER));
         try {
             String resultAttr = ATTR_ERROR;
-            if (ORDER_SERVICE.deleteOrder(orderId)) {
-                resultAttr = ATTR_SUCCESS;
+            if (user.getRole().equals(UserRole.CLIENT)) {
+                if (ORDER_SERVICE.deleteOrder(orderId)) {
+                    resultAttr = ATTR_SUCCESS;
+                }
+                request.getSession().setAttribute(resultAttr, ActionResult.ORDER_DELETED);
+                page = COMMAND_VIEW_ORDERS;
+            } else {
+                if (ORDER_SERVICE.deleteLateOrder(orderId)) {
+                    resultAttr = ATTR_SUCCESS;
+                }
+                request.getSession().setAttribute(resultAttr, ActionResult.LATE_ORDER_DELETED);
+                page = COMMAND_VIEW_UNPAID_ORDERS;
             }
-            request.getSession().setAttribute(resultAttr, ActionResult.ORDER_DELETED);
-            page=COMMAND_VIEW_ORDERS;
+
         } catch (ServiceException e) {
             throw new CommandException(e);
         }
